refactor: relocate dto
