+++ src/main/java/com/scg/stop/aihub/dto/request/AiHubModelRequest.java
@@ -1,4 +1,4 @@
-package com.scg.stop.aihub.dto;
+package com.scg.stop.aihub.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,12 +9,12 @@ import java.util.List;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class AiHubDatasetRequest {
+public class AiHubModelRequest {
 
     private String title; // 제목 : text
-    private List<String> dataTypes; // 데이터 유형 : multiselect
+    private List<String> learningModels;  // 학습 모델 : multiselect
     private List<String> topics;  // 주제 분류 : multiselect
-    private List<Integer> developmentYears;  // 구축 년도 : multiselect
+    private List<Integer> developmentYears;  // 개발 년도 : multiselect
     private String professor; // 담당 교수 : text
     private List<String> participants; // 참여 학생 : text (comma separated)
 }
