+++ src/main/java/com/scg/stop/aihub/dto/response/AiHubDatasetResponse.java
@@ -1,4 +1,4 @@
-package com.scg.stop.aihub.dto;
+package com.scg.stop.aihub.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
